% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_.R
\name{cut_}
\alias{cut_}
\title{Divide Numeric Into Intervals}
\usage{
cut_(
  x,
  breaks = quantile(x, probs = probs, na.rm = TRUE),
  probs,
  data.name = substitute(x),
  fmt = "\%.3g",
  ...
)
}
\arguments{
\item{x}{\link[base]{integer}, \link[base]{numeric}, \link[base]{difftime}, \link[base]{Date},
\link[base]{POSIXct} \link[base]{vector} or \link[base]{matrix}.}

\item{breaks}{\link[base]{vector} of the same \link[base]{class} as \code{x}.
\code{-Inf} and \code{Inf} will be automatically added}

\item{probs}{\link[base]{double} \link[base]{vector} from 0 to 1,
probabilities to specify the \link[stats]{quantile}s to be used as \code{breaks}}

\item{data.name}{\link[base]{character} scalar, name of data.
R \link[base]{language} is also accepted.
Default is the argument call of \code{x}}

\item{fmt}{see functions \code{\link[=cut_levels]{cut_levels()}} and \link[base]{sprintf}}

\item{...}{additional parameters \code{right} and \code{include.lowest} of
functions \link[base]{cut.default} and \link[base]{.bincode}.}
}
\description{
Divide numeric into intervals; an alternative to function \link[base]{cut.default}.
}
\details{
Function \code{\link[=cut_]{cut_()}} is different from function \link[base]{cut.default}, that
\itemize{
\item {More classes of \code{x} are accepted, see \strong{Arguments}}
\item {\code{-Inf} and \code{Inf} are added to \code{breaks},
so that the values outside of \code{breaks} will be correctly categorized,
instead of returning an \code{NA_integer} per \link[base]{.bincode}}
\item {More user-friendly \link[base]{factor} \link[base]{levels}, see helper function \code{\link[=cut_levels]{cut_levels()}}}
}
}
\examples{
datasets::morley$Speed |> 
 cut_(breaks = 100*(8:9), data.name = 'speed') |> 
 table()
datasets::morley$Speed |> 
 cut_(probs = c(.3, .6)) |> 
 table()

datasets::VADeaths |> 
 cut_(probs = c(.3, .6), data.name = '\U1F480', fmt = '\%.3g\u2030')

datasets::JohnsonJohnson |> 
 zoo::as.Date.ts() |>
 cut_(breaks = as.Date(c('1970-01-01', '1975-01-01')), 
   data.name = 'JnJ') |>
 table()

}
\keyword{internal}
