% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_.R
\name{cut_levels}
\alias{cut_levels}
\title{\code{\link[=cut_levels]{cut_levels()}}}
\usage{
cut_levels(breaks, right = TRUE, include.lowest = TRUE, data.name = "x")
}
\arguments{
\item{breaks}{\link[base]{vector} of the same \link[base]{class} as \code{x}.
\code{-Inf} and \code{Inf} will be automatically added}

\item{right}{\link[base]{logical} scalar, default \code{TRUE},
see functions \link[base]{cut.default} and \link[base]{.bincode}.}

\item{include.lowest}{\link[base]{logical} scalar, default \code{TRUE},
see functions \link[base]{cut.default} and \link[base]{.bincode}.}

\item{data.name}{\link[base]{character} scalar, name of data.
R \link[base]{language} is also accepted.
Default is the argument call of \code{x}}
}
\description{
\code{\link[=cut_levels]{cut_levels()}}
}
\examples{
## Examples on Helper Function cut_levels()
foo = \(...) {
 cbind(
  numeric(0) |> cut.default(...) |> levels(),
  cut_levels(...)
 )
}
foo(breaks = 1:4, right = TRUE, include.lowest = TRUE)
foo(breaks = 1:4, right = FALSE, include.lowest = TRUE)
foo(breaks = 1:4, right = TRUE, include.lowest = FALSE)
foo(breaks = 1:4, right = FALSE, include.lowest = FALSE)
set.seed(29); foo(breaks = c(-Inf, sort(rnorm(1:3)), Inf))
}
\keyword{internal}
