% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_set.R
\name{factor-set}
\alias{factor-set}
\alias{factor<-}
\alias{ordered<-}
\title{Syntactic Sugar to Set (\link[base]{ordered}) \link[base]{factor} \link[base]{levels}}
\usage{
factor(x, plus = 0L, ordered = FALSE) <- value

ordered(x, plus = 0L) <- value
}
\arguments{
\item{x}{\link[base]{integer} \link[base]{vector},
dummy coding indices}

\item{plus}{\link[base]{integer} scalar,
value to be added to \code{x}, default \code{0L}.
If the dummy coding starts from \code{0L} (instead of from \code{1L}), then use \code{plus = 1L}}

\item{ordered}{\link[base]{logical} scalar, should the returned \link[base]{factor} be
\link[base]{ordered} (\code{TRUE}) or un-ordered (\code{FALSE}, default)}

\item{value}{\link[base]{character} \link[base]{vector},
\link[base]{levels} of output \link[base]{factor}}
}
\value{
Syntactic sugar \link{factor<-} returns a \link[base]{factor}.

Syntactic sugar \link{ordered<-} returns an \link[base]{ordered} \link[base]{factor}.
}
\description{
Syntactic sugar to set (\link[base]{ordered}) \link[base]{factor} \link[base]{levels}.
}
\examples{
set.seed(22); (x1 = x2 = sample.int(n = 5L, size = 20, replace = TRUE))
factor(x1) = letters[1:5]; x1
ordered(x2) = LETTERS[1:5]; x2

set.seed(14); (x10 = x20 = sample.int(n = 5L, size = 20, replace = TRUE) - 1L)
factor(x10, plus = 1L) = letters[1:5]; x10
ordered(x20, plus = 1L) = LETTERS[1:5]; x20

# some exceptions
x = 1:4
factor(x) = c('a', 'b', 'c', 'c')
x # duplicated levels dropped

x = 1:4
factor(x) = c('a', 'b', NA_character_, 'c')
x # missing level converted to missing entry

set.seed(32); (x = array(sample.int(4, size = 20, replace = TRUE), dim = c(4,5)))
factor(x) = letters[1:4]
x # respects other attributes
}
\keyword{internal}
