% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{factor.tzh-package}
\alias{factor.tzh}
\alias{factor.tzh-package}
\title{factor.tzh: Tingting's extension of factor object}
\description{
Tingting's extension of factor object.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
