% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode2factor.R
\name{recode2factor}
\alias{recode2factor}
\title{\link[base]{factor} Return for \link[dplyr]{recode_values}}
\usage{
recode2factor(x, ..., default = NULL, ptype = NULL, envir = parent.frame())
}
\arguments{
\item{x, ..., default, ptype}{see parameters of function \link[dplyr]{recode_values}}

\item{envir}{\link[base]{environment} to \link[base]{eval}uate function \link[dplyr]{recode_values}.
Default \code{parent.frame()}}
}
\value{
Function \code{\link[=recode2factor]{recode2factor()}} returns a \link[base]{factor}.
}
\description{
..
}
\details{
If argument \code{default} is missing, function \code{\link[=recode2factor]{recode2factor()}} converts
the return of function \link[dplyr]{recode_values} into a \link[base]{factor}.
The order of \link[base]{levels} follows the order of formulas in
\link[rlang:dyn-dots]{dynamic dots} \code{...}.
}
\note{
Functions \link[dplyr]{case_match} and \link[dplyr]{recode_factor}
are deprecated (since v1.2.0) with \link[dplyr]{recode_values}.
}
\examples{
# from ?dplyr::recode_values
x = c('a', 'b', 'a', 'd', 'b', NA, 'c', 'e')
dplyr::recode_values(x, 'a' ~ 1, 'b' ~ 2, 'c' ~ 3, 'd' ~ 4)
recode2factor(x, 'a' ~ 1, 'b' ~ 2, 'c' ~ 3, 'd' ~ 4)
recode2factor(x, 'a' ~ 1, 'b' ~ 2, 'c' ~ 1, 'd' ~ 2)

recode2factor(x, 'a' ~ 1, 'd' ~ 4, 'b' ~ 2, 'c' ~ 3) # order matters!

y = c(1, 2, 1, 3, 1, NA, 2, 4)
recode2factor(y, NA ~ 0, default = y)
}
\keyword{internal}
