% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.factor.R
\name{Summary.factor}
\alias{Summary.factor}
\title{\link[base]{max} and \link[base]{min} of \link[base]{factor}}
\usage{
\method{Summary}{factor}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{a \link[base]{factor} object}

\item{...}{currently not used!!}

\item{na.rm}{\link[base]{logical} scalar}
}
\value{
Function \code{\link[=Summary.factor]{Summary.factor()}}
overwrites function \link[base]{Summary.factor} and
returns a \link[base]{factor}.
}
\description{
..
}
\examples{
(x = c(NA_integer_, sample.int(3L, size = 10L, replace = TRUE)) |>
 structure(levels = letters[1:3], class = 'factor'))
max(x, na.rm = FALSE)
max(x, na.rm = TRUE)
min(x, na.rm = TRUE)

(x0 = rep(NA_integer_, times = 20L) |>
 structure(levels = letters[1:3], class = 'factor'))
max(x0)
min(x0)
}
\keyword{internal}
