% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_.R
\name{cut_}
\alias{cut_}
\alias{cut_levels}
\title{Divide Numeric Into Intervals}
\usage{
cut_(
  x,
  breaks = quantile(x, probs = break_probs, na.rm = TRUE),
  break_probs,
  right = TRUE,
  include.lowest = TRUE,
  data.name = substitute(x),
  ...
)

cut_levels(breaks, right = TRUE, include.lowest = TRUE, data.name = "x")
}
\arguments{
\item{x}{\link[base]{integer}, \link[base]{numeric}, \link[base]{difftime}, \link[base]{Date},
\link[base]{POSIXct} \link[base]{vector} or \link[base]{matrix}.}

\item{breaks}{\link[base]{vector} of the same \link[base]{class} as \code{x}.
\code{-Inf} and \code{Inf} will be automatically added}

\item{break_probs}{\link[base]{double} \link[base]{vector} from 0 to 1,
probabilities to specify the \link[stats]{quantile}s to be used as \code{breaks}}

\item{right}{\link[base]{logical} scalar, default \code{TRUE},
see functions \link[base]{cut.default} and \link[base]{.bincode}.}

\item{include.lowest}{\link[base]{logical} scalar, default \code{TRUE},
see functions \link[base]{cut.default} and \link[base]{.bincode}.}

\item{data.name}{\link[base]{character} scalar, name of data.
R \link[base]{language} is also accepted.
Default is the argument call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\description{
Divide numeric into intervals; an alternative to function \link[base]{cut.default}.
}
\details{
Function \code{\link[=cut_]{cut_()}} is different from function \link[base]{cut.default}, that
\itemize{
\item {More classes of \code{x} are accepted, see \strong{Arguments}}
\item {\code{-Inf} and \code{Inf} are added to \code{breaks},
so that the values outside of \code{breaks} will be correctly categorized,
instead of returning an \code{NA_integer} per \link[base]{.bincode}}
\item {More user-friendly \link[base]{factor} \link[base]{levels}, see helper function \link{cut_levels}}
}
}
\examples{
morley$Speed |> cut_(breaks = 100*(8:9), data.name = 'speed') |> table()
morley$Speed |> cut_(break_probs = c(.3, .6)) |> table()

VADeaths |> cut_(break_probs = c(.3, .6))

JohnsonJohnson |> 
 zoo::as.Date.ts() |>
 cut_(breaks = as.Date(c('1970-01-01', '1975-01-01')), 
   data.name = 'JnJ') |>
 table()

## Examples on Helper Function cut_levels()
foo = function(...) cbind(
 cut.default(numeric(0), ...) |> levels(),
 cut_levels(...)
)
foo(breaks = 1:4, right = TRUE, include.lowest = TRUE)
foo(breaks = 1:4, right = FALSE, include.lowest = TRUE)
foo(breaks = 1:4, right = TRUE, include.lowest = FALSE)
foo(breaks = 1:4, right = FALSE, include.lowest = FALSE)
set.seed(29); foo(breaks = c(-Inf, sort(rnorm(1:3)), Inf))
}
\keyword{internal}
