% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_.R
\name{cut_levels}
\alias{cut_levels}
\title{\code{\link[=cut_levels]{cut_levels()}}}
\usage{
cut_levels(
  breaks,
  right = TRUE,
  include.lowest = TRUE,
  fmt = "\%.3g",
  data.name = "x"
)
}
\arguments{
\item{breaks, right, include.lowest}{see functions \link[base]{cut.default} and \link[base]{.bincode}.}

\item{fmt}{see function \link[base]{sprintf}}

\item{data.name}{\link[base]{character} scalar or \link[base]{language}, see function \code{\link[=cut_]{cut_()}}}
}
\description{
\code{\link[=cut_levels]{cut_levels()}}
}
\examples{
foo = \(...) {
 cbind(
  numeric(0) |> cut.default(...) |> levels(),
  cut_levels(...)
 )
}
foo(breaks = c(-Inf, 1:4, Inf), right = TRUE, include.lowest = TRUE)
foo(breaks = c(-Inf, 1:4, Inf), right = FALSE, include.lowest = TRUE)
foo(breaks = c(-Inf, 1:4, Inf), right = TRUE, include.lowest = FALSE)
foo(breaks = c(-Inf, 1:4, Inf), right = FALSE, include.lowest = FALSE)
}
\keyword{internal}
